[gem] rename to sidekiq-portal
