+++ .travis.yml
@@ -15,6 +15,3 @@ matrix:
   - rvm: jruby-head
 script:
 - bundle exec rspec
-notifications:
-  slack:
-    secure: G4xCbE/XyJnmr46fS6mItvEfnGZ7QpJuXnnjBe/D7lYFAZ3RfhfV/opOpd9cqp0WkptAEEiW9FdvC7hTsq8wFfmdJgmuuZqNTB3R4hmSyg09NKdwsu/VrXDl9Cpm0UWmTetC2nMqw52jUU4VoeZttvwAQaEtKE2rBE9HeEjU6LRhy8VUo1KDBMl2DqAsa6Khzc2vJUjtU+fUrSQsGLKHcd+pCAA9FGx3PCZ06oImugoNpQ1auBs4kci1QDaXArP11udne8q4R7mYnhg0ASKHSggZKHdRt3zGQpVGHPiZ9IJs9Op5meaCKu9lTFqSVUkEamKuUL2vfg0wqawlOQ+2rPwOQTUFZkYe0RbU6jBFKo3VGf4oF/sJoLrLD+CeczwbJeNIg9QE36rnFbm0vQV9eRZhJn2190n7z8EuQP9vucLPiY2jbtI4BE2DEMNqTkrVxOw1PbjGxBQiQ/sDRrjHCBvIWrCYsV8Zqwk/fZRhZdvFizwwQCUu/yca6kOM+BD55Pkt3BsF4lJec06J/mAoShZqhXx7cKvCC7wMRDn2d/JWMWVz3sGB0CDeuN4HpaiNkg1ln95tCqHD1UrV15LUz/g6uTYtWKlMGaVKXFwCZzd0nzBSDiPiH7NzHqsw9e3s8LbxOjtNoIc+VElDTW3ACwPD6KZ5uJC76k+YIpkcOxc=
