+++ README.md
@@ -1,4 +1,4 @@
-# sidekiq-scheduler-mock &middot; [![Build Status](https://travis-ci.org/0exp/sidekiq-scheduler-mock.svg?branch=master)](https://travis-ci.org/0exp/sidekiq-scheduler-mock) [![Coverage Status](https://coveralls.io/repos/github/0exp/sidekiq-scheduler-mock/badge.svg?branch=master)](https://coveralls.io/github/0exp/sidekiq-scheduler-mock?branch=master)
+# sidekiq-portal
 
 In active development...
