+++ lib/sidekiq-portal.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative 'sidekiq_portal'
