+++ lib/sidekiq_portal/version.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Sidekiq
+  class Portal
+    # @return [String]
+    #
+    # @api public
+    # @since 0.0.0
+    VERSION = '0.0.0'
+  end
+end
