--- lib/sidekiq_scheduler_mock/extensions/scheduler_aliases.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Sidekiq
-  # @api public
-  # @since 0.1.0
-  SchedulerMock = SidekiqSchedulerMock
-
-  class Testing
-    # @api public
-    # @since 0.1.0
-    SchedulerMock = SidekiqSchedulerMock
-  end
-end
