--- lib/sidekiq_scheduler_mock/version.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class SidekiqSchedulerMock
-  # @return [String]
-  #
-  # @api public
-  # @since 0.1.0
-  VERSION = '0.1.0'
-end
