+++ sidekiq-portal.gemspec
@@ -2,21 +2,21 @@
 
 lib = File.expand_path('lib', __dir__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'sidekiq_scheduler_mock/version'
+require 'sidekiq_portal/version'
 
 Gem::Specification.new do |spec|
   spec.required_ruby_version = '>= 2.3.8'
 
-  spec.name          = 'sidekiq-scheduler-mock'
-  spec.version       = SidekiqSchedulerMock::VERSION
+  spec.name          = 'sidekiq-portal'
+  spec.version       = Sidekiq::Portal::VERSION
   spec.authors       = ['Rustam Ibragimov']
   spec.email         = ['iamdaiver@gmail.com']
 
-  spec.summary       = 'SidekiqScheduler mock for your tests.'
-  spec.description   = 'SidekiqScheduler mock for your tests that emulates ' \
-                       'main functionality of the real SidekiqScheduler gem.'
+  spec.summary       = 'Sidekiq::Portal - worker scheduler mock for your tests.'
+  spec.description   = 'Sidekiq::Portal - worker scheduler mock for your tests that emulates ' \
+                       'job activity when you work with time traveling.'
 
-  spec.homepage      = 'https://github.com/0exp/sidekiq-scheduler-mock'
+  spec.homepage      = 'https://github.com/0exp/sidekiq-portal'
   spec.license       = 'MIT'
   spec.bindir        = 'bin'
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
