+++ spec/smoke_spec.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-describe 'Load Scheduler Mock' do
-  specify { SidekiqSchedulerMock.setup! }
+describe 'Load Sidekiq Portal' do
+  specify { Sidekiq::Portal.setup! }
 end
